Cleanup offsetof()
cppreference says the expression's type should be size_t.  I noticed
when passing to an OpenGL function that takes a void* that I didn't have
to cast the result for the first struct field, which seemed like an odd
inconsistency.

+++ core.h
@@ -46,7 +46,7 @@ typedef struct timespec timepoint_t;
 #define B2PC(x) (x), countof(x)
 
 #ifndef offsetof
-#define offsetof(s, m) (&(((s*)(NULL))->m))
+#define offsetof(s, m) ((size_t)&(((s*)(NULL))->m))
 #endif
 
 #define polarity(x) (((x) > 0) - ((x) < 0))
